Summarize the results from lcvp_search

Summarizes the matching results from lcvp_search.

lcvp_summary(lcvp_match)

Arguments

lcvp_match

The output data.frame from lcvp_search.

Value

It returns:

The number of species searched;

The number and percentage of species names found in the LCVP data;

The number and percentage of species names exactly matched;

The number and percentage of species names fuzzy matched;

The number and percentage of author names exactly matched;

The number and percentage of infracategories exactly matched.

Details

lcvp_summary gives a report on the searching results using the lcvp_search function. Indicating the number of species searched, how many were matched. Among the matched names, it indicates how many were exactly or fuzzy matched. Then it checks how many author and infracategory names were exactly matched. Note that if authors or infracategory is not provided, it will be considered a no match.

References

Freiberg, M., Winter, M., Gentile, A. et al. LCVP, The Leipzig catalogue of vascular plants, a new taxonomic reference list for all known vascular plants. Sci Data 7, 416 (2020). https://doi.org/10.1038/s41597-020-00702-z

See also

Author

Bruno Vilela & Alexander Ziska

Examples

# Ensure that LCVP package is available before running the example.
# If it is not, see the `lcvplants` package vignette for details
# on installing the required data package.
if (requireNamespace("LCVP", quietly = TRUE)) { # Do not run this

# Perform the search
x <- lcvp_search(c("Hibiscus abelmoschus var. betulifolius Mast.",
"Hibiscus abutiloides Willd.",
"Hibiscus aculeatus",
"Hibiscus acuminatis",
"Hibiscus error"), 
max_distance = 1)

# Summarize the results
lcvp_summary(x)

}