The Leipzig Catalogue of Vascular Plants (LCVP) - An Improved Taxonomic Reference List for All Known Vascular Plants

A package for large-scale taxonomic harmonization of plant names by fuzzy matching and synonymy resolution against the Leipzig Plant Catalogue as taxonomic backbone. Submission of single names or list of species names is possible. The Leipzig Plant Catalogue is an updated taxonomic backbone based on the plant list comprising more than 1,300,000 names and 350,000 accepted names (this data is in the LCVP package).

lcvplants functions

lcvp_search: Standardize plant names according to the LCVP

lcvp_summary: Summarize the results from lcvp_search

lcvp_group_search: Search plants by taxa or author names in the LCVP

lcvp_fuzzy_search: Fuzzy match plant names according to the LCVP

lcvp_join: Join two data.frames using the LCVP

lcvp_match: Match two name lists using the LCVP

lcvp_solve_dups: Solve duplicated names by summarizing traits

References

The Leipzig Catalogue of Vascular Plants (LCVP) - An improved taxonomic reference list for all known vascular plants.

See also

Author

Bruno Vilela Alessandro Gentile Martin Freiberg Marten Winter Alexander Zizka

Maintainer: Bruno Vilela bvilela.bv@gmail.com