Match two name lists using the Leipzig Catalogue of Plants (LCVP)

Matches and compares two name lists based on the taxonomic resolution of plant taxa names listed in the "Leipzig Catalogue of Vascular Plants" (LCVP).

lcvp_match(
  splist1,
  splist2,
  max.distance = 0.1,
  include_all = TRUE,
  identify_dups = TRUE
)

Arguments

splist1
A character vector specifying the reference input taxon to be matched. Each element including genus and specific epithet and, potentially, infraspecific rank, infraspecific name, and author name. Only valid characters are allowed (see base:validEnc).

splist2
A character vector specifying the input taxon to match splist1. Each element including genus and specific epithet and, potentially, infraspecific rank, infraspecific name, and author name. Only valid characters are allowed (see base:validEnc).

max.distance
It represents the maximum distance allowed for a match when comparing the submitted name with the closest name matches in the LCVP. Expressed either as integer, or as a fraction of the pattern length times the maximal transformation cost (will be replaced by the smallest integer not less than the corresponding fraction). See agrep for more details.

include_all
If TRUE (default), it will include all species in both splist1 and splist2. If FALSE, it will exclude species only found in splist2.

identify_dups
If TRUE (default), a column indicating the position of duplicated LCVP output names in the resulting data.frame.

Value

A data.frame with the following columns:

Species.List.1: Taxa name list provided by the user in the splist1.

Species.List.2: Taxa name list provided by the user in the splist2.

LCVP.Input.Taxon: Matched taxa names listed in the LCVP data.

Status: Nomenclature status: 'accepted', 'synonym', 'unresolved' or 'external'.

LCVP.Output.Taxon: The list of the accepted plant taxa names according to the LCVP.

Match.Position.2to1: positions of the names in splist1 in splist2. Can be used to reorder splist2 to match splist1.

Duplicated.Output.Position: If identify_dups = TRUE, it indicates the position of duplicated names in LCVP.Output.Taxon column. This may occur if two inputs are now synonyms. It will output NA if there is no duplicated for the species name.

See LCVP:tab_lcvp for more details.

If include_all = TRUE, all species will be included. Ordered based on the splist1, and followed by non-matched names in splist2. If include_all = FALSE, non-matched names in splist2 are not included.

References

Freiberg, M., Winter, M., Gentile, A. et al. LCVP, The Leipzig catalogue of vascular plants, a new taxonomic reference list for all known vascular plants. Sci Data 7, 416 (2020). https://doi.org/10.1038/s41597-020-00702-z

See also

Author

Bruno Vilela & Alexander Ziska

Examples

# Ensure that LCVP package is available before running the example.
# If it is not, see the `lcvplants` package vignette for details
# on installing the required data package.
if (requireNamespace("LCVP", quietly = TRUE)) { # Do not run this

# Generate two lists of species name
splist1 <- sample(LCVP::tab_lcvp$Input.Taxon[5:100])
splist2 <- sample(LCVP::tab_lcvp$Input.Taxon[100:2])

# Including all species in both lists
lcvp_match(splist1, splist2, include_all = TRUE)

# Including all species only in the first list
matchLists <- lcvp_match(splist1, splist2, include_all = FALSE)
## This can be used to quickly change positions in splist2 to match splist1
splist2[matchLists$Match.Position.2to1]

}